Refactor chart options
